Pulled models into single package, renamed other modules, fixed imports
