+++ models/utils.py
@@ -1,0 +1,7 @@
+"""
+Functions common to the execution of all models in the project.
+
+Written by Michael Wheeler and Jay Sherman.
+"""
+
+
