+++ text_reports/utils.py
